Break up into lms_type_code into lmsTypecode and lmotsTypecode
This commit replaces lms_type_code (which was used for typecodes for
both LMS and LM-OTS) with two types: lmsTypecode (for LMS typecodes) and
lmotsTypecode (for LM-OTS typecodes).

While the typecodes for both LMS and LM-OTS parameter sets are both
uint32 values, and the original sets in RFC8554 didn't overlap, the
namespaces are distinct: see
https://www.iana.org/assignments/leighton-micali-signatures/leighton-micali-signatures.xhtml

Separating the types also allows greater clarity of code: the expected
type of a typecode is explicit now, and type safety will reduce future
bugs.

The new type names use camelCase instead of snake_case. This is
idiomatic for Go:
https://google.github.io/styleguide/go/decisions#underscores

The original choice not to export lms_type_code is respected by this
change. Not making the type public is an uncommon choice, but it does
allow internal refactors like this one to avoid breaking callers.